# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gerir.me API",
    "description": "API REST para o sistema Gerir.me - Gerenciamento de Despesas Pessoais",
    "version": "1.0.0",
    "contact": {
      "name": "Edcleryton",
      "email": "contato@gerirme.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Expense": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único da despesa"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "ID do usuário proprietário"
          },
          "description": {
            "type": "string",
            "minLength": 3,
            "description": "Descrição da despesa"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01,
            "description": "Valor da despesa"
          },
          "category": {
            "type": "string",
            "minLength": 2,
            "description": "Categoria da despesa"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "description": "Data de vencimento"
          },
          "isPaid": {
            "type": "boolean",
            "description": "Status de pagamento"
          },
          "paidDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Data do pagamento"
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Se a despesa é recorrente"
          },
          "recurrenceType": {
            "type": "string",
            "enum": ["weekly", "monthly", "yearly"],
            "nullable": true,
            "description": "Tipo de recorrência"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags da despesa"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionais"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data da última atualização"
          }
        }
      },
...
      "CreateExpenseRequest": {
        "type": "object",
        "required": ["description", "amount", "category", "dueDate"],
        "properties": {
          "description": {
            "type": "string",
            "minLength": 3,
            "description": "Descrição da despesa"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01,
            "description": "Valor da despesa"
          },
          "category": {
            "type": "string",
            "minLength": 2,
            "description": "Categoria da despesa"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "description": "Data de vencimento"
          },
          "isRecurring": {
            "type": "boolean",
            "default": false,
            "description": "Se a despesa é recorrente"
          },
          "recurrenceType": {
            "type": "string",
            "enum": ["weekly", "monthly", "yearly"],
            "description": "Tipo de recorrência (obrigatório se isRecurring=true)"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags da despesa"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionais"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Tipo do erro"
          },
          "message": {
            "type": "string",
            "description": "Mensagem de erro"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Lista de erros de validação"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/expenses": {
      "get": {
        "tags": ["Expenses"],
        "summary": "Listar despesas",
        "description": "Retorna todas as despesas do usuário autenticado com filtros opcionais",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filtrar por categoria",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status",
            "schema": {
              "type": "string",
              "enum": ["paid", "unpaid", "overdue"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de despesas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "expenses": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Expense"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "totalItems": {
                          "type": "integer"
                        },
                        "itemsPerPage": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Expenses"],
        "summary": "Criar nova despesa",
        "description": "Cria uma nova despesa para o usuário autenticado",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Despesa criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "expense": {
                      "$ref": "#/components/schemas/Expense"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}